Create BaseFileReader that readers (e.g. the ParquetReader) should inherit from
